comment about fold made obsolete by Foldable1
I wasn't planning on implementing fold functions since they can be
derived from the traversable functions, but it does make sense for a
separate Foldable1 type class to exist, because even though all of our
shapes are Traversable1, there might be other data structures which are
Foldable1 but not Traversable1. So it makes sense to also implement
Foldable1, if only by calling a generic function which reuses the
Traversable1 instance.