towards the Structured shape
Lists, Trees, and Tries, all of those type-level indices assume a finite
number of possible fields. This won't work if we want to point into a
Map. Paths should work with an infinite number of fields, but for that
we must untangle them from the Lists, Trees, and Tries they are indexing
into. An associated class would allow us to talk both about paths into
Lists, Trees, Tries, and Maps.