vinyl: more an inspiration than a building block
The original plan was to add shapes like RecTree on top of vinyl's Rec,
but that doesn't work for types with a potentially-infinite set of
fields, like Map. A new approach is needed, and I think we need to make
a clean break from vinyl by keeping the 'f' idea but not the type-level
list/tree idea.