redesign based on Monoid
Instead of a single type class which does both patch composition,
validation, and inversion, let's divided it into three smaller pieces.
If we're only concerned about composition, we can simply delegate most
of the work to the Monoid type class.

Note that I have updated the implementation, but not yet the
documentation, so the haddocks currently say the wrong thing.