downgrade Prisms to Reviews
The forward direction of those Prisms is misleading, it might look
like _PatchLeft gets the left part of a PatchEither, when in fact it
only does so when the right part is mempty. A Review only supports
the backward direction, and that's the only direction I envision being
useful anyway.