embrace Language.Haskell.TH.Lib's helpers
the convention in that module, which we adopt in ours in order to
simplify the code, is to keep Q computations wrapped in Q, instead of
running them in Q in order to obtain the wrapped values. This means that
many of our datatypes lose their Show and Eq instances.